requirements.txt
psycopg2
kafka-python

aiopg
aiohttp
aiokafka
